requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
pypdf>=3.17.0
pytesseract>=0.3.10
pdfplumber>=0.10.0
ebooklib>=0.18

# NLP & ML
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.6.0
sentencepiece>=0.1.99
spacy>=3.7.0
transformers[torch]>=4.36.0

# Vector Database
qdrant-client>=1.9.0
grpcio>=1.60.0

# LLM
google-generativeai>=0.4.0
anthropic>=0.18.0
openai>=1.10.0
cohere>=5.5.0

# OSINT
playwright>=1.40.0
scrapy>=2.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
dnspython>=2.4.0
shodan>=1.30.0

# Orchestration & Agents
crewai>=0.28.0
autogen>=0.2.0
taskloop>=0.1.9

# Monitoring & Logging
prometheus-client>=0.19.0
grafana-api>=1.1.0
sentry-sdk>=1.40.0
python-dotenv>=1.0.0
loguru>=0.7.0
